pyproject.toml
[tool.poetry]
name = "abz-test-task"
version = "0.1.0"
description = "catalog"
authors = ["AleksandrKosmylev <alex.kosmylev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-mptt = "^0.16.0"
factory-boy = "^3.3.0"
django-bootstrap5 = "^24.2"
psycopg2 = "^2.9.9"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
